requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
future==0.17.1
numpy==1.17.0
pandas==0.24.2
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.3
python-telegram-bot==11.1.0
pytz==2019.2
redis==3.3.8
six==1.12.0
SQLAlchemy==1.3.7
